requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
asyncio-extras==1.3.2
asyncpraw==7.5.0
asyncprawcore==2.3.0
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.1.3
colorama==0.4.4
discord.py==1.7.3
docutils==0.18.1
ffmpeg==1.4
Flask==1.1.1
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==5.2.0
mutagen==1.45.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
praw==7.6.0
prawcore==2.3.0
protobuf==3.19.3
psutil==5.9.0
psycopg2==2.9.3
py==1.11.0
py-cord==1.7.3
pycparser==2.21
pycryptodomex==3.12.0
pyfiglet==0.7
Pygments==2.11.2
PyNaCl==1.4.0
pyparsing==3.0.7
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-forecastio==1.4.0
requests==2.27.1
responses==0.20.0
rule34==1.8.1
semantic-version==2.9.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
spotipy==2.19.0
toml==0.10.2
typing_extensions==4.0.1
update-checker==0.18.0
urllib3==1.26.8
vcrpy==4.1.1
websocket-client==1.3.2
websockets==10.1
Werkzeug==2.1.2
wrapt==1.13.3
yarl==1.7.2
youtube-dl==2021.12.17
yt-dlp==2021.12.27